+++ src/components/hero/campaignHero.tsx
@@ -4,9 +4,16 @@ import { useEffect, useState } from "react";
 import CampaignHeroSubText from "./campaignHeroSubText";
 import CampaignHeroHeading from "./campaignHeroHeading";
 import CampaignHeroLoading from "./campaignHeroLoading";
+import defaultFeaturedItem from "../../data/defaultFeaturedItemData";
 
-const CampaignHero = ({ imageID }) => {
-  const [data, setData] = useState<any>({});
+import appConfig from "appConfig";
+import { FeaturedItemData } from "@/types/FeaturedItemData";
+
+const CampaignHero = () => {
+  const defaultFeaturedItemResponse =
+    defaultFeaturedItem[appConfig["environment"]];
+
+  const [data, setData] = useState<FeaturedItemData>();
 
   useEffect(() => {
     const fetchData = async () => {
@@ -16,7 +23,13 @@ const CampaignHero = ({ imageID }) => {
     };
 
     fetchData();
-  }, [imageID]);
+  }, []);
+
+  const handleError = (e: any) => {
+    console.log(e);
+    setData(defaultFeaturedItemResponse);
+    console.log("data is:", data);
+  };
 
   return data?.featuredItem ? (
     <Hero
@@ -24,6 +37,7 @@ const CampaignHero = ({ imageID }) => {
       backgroundColor="ui.bg.default"
       isDarkBackgroundImage
       heroType="campaign"
+      isDarkText
       heading={
         <CampaignHeroHeading
           numberOfDigitizedItems={data.numberOfDigitizedItems}
@@ -32,6 +46,10 @@ const CampaignHero = ({ imageID }) => {
       imageProps={{
         alt: data.featuredItem.title,
         src: data.featuredItem.foregroundImageSrc,
+        fallbackSrc: data.featuredItem.foregroundImageSrc,
+        onError: (_event) => {
+          handleError(_event);
+        },
       }}
       subHeaderText={<CampaignHeroSubText featuredItem={data.featuredItem} />}
     />